Correctly define TT_OS2 struct.
The existing definition was missing some fields after ulUnicodeRange4
and before ulCodePageRange1 and defined fsType as an FT_Short instead of
as an FT_UShort. This would have meant that some fields would not have
been read correctly. Add the missing fields.

A corresponding change to PistonDevelopers/freetype-rs adds a test.