Build freetype from source on android
This adapts the build process of servo-freetype-sys to build freetype2
from source when on the android target.

Since this is the only difference preventing servo from switching to
freetype-sys, it would then be possible to deprecate the
servo-freetype-sys package in favor of having only a single library
linking to the freetype project.

Without unifying the different freetype bindings, it is not possible for
the Alacritty project to build without having to manually replace one of
the depedencies using the patch section in the Cargo.toml, which is not
only ugly but also prevents pushing to crates.io. Of course it would be
possible to completely switch to servo-freetype, but I think it would be
beneficial for the entire Rust ecosystem to be able to freely combine
the different libraries.

This was originally discussed with servo in
https://github.com/servo/libfontconfig/pull/46, where they've offered to
cooperate on this effort. While they've stated that they're willing to
give up control over freetype-sys completely, it might be nice to share
control over the repository and crates.io.