Export C include path information into Cargo
Regardless of whether we build FreeType locally or not, print a magic
string that depending crates can use to find the path to the FreeType C
includes. That way, if those crates have their own C code that needs to
interface with FreeType, there will be an easy and reliable way to do
so.