Add a `bundled` Cargo feature
Previously, `freetype-sys` would first try to find the system FreeType
via pkg-config and then fall back to compiling the vendored version of
FreeType unconditionally.

This change adds a `bundled` feature which allows the users of the crate
to have more control over whether or not the static version of FreeType
is used. If the feature is not passed, FreeType will not be
statically-linked even if it isn't found via pkg-config. This matches
the approach that `harfbuzz-sys` uses as well as `openssl-sys` (via the
`vendored` feature).